requirements.txt
python-telegram-bot==20.3
requests
Flask==2.2.2
gunicorn==20.1.0